HEE-275: (1) Split 'app/assets/javascript/main.js' up into component specific js files and import, (2) Update gulp to compile main.js and its minified version using webpack, (3) Update gulp 'bundle' task to bundle only minified CSS and JS assets.
