+++ app/assets/components/nav-map/nav-map.js
@@ -1,21 +1,21 @@
-/**
-* Map
-* Elements with the selector '.nhsuk-region' are passed into this class
-*/
-
-class NavMap {
+export default () => {
+  /**
+  * Map
+  * Elements with the selector '.nhsuk-region' are passed into this class
+  */
+  class NavMap {
     constructor(map, svg) {
       this.map = map;
       this.svg = svg;
       this.regions = [...svg.getElementsByClassName('nhsuk-region')];
       this.list = [...map.querySelectorAll('#regionList li a')];
-  
+
       this.addLinksToMap();
       this.addClasses();
       this.mapEventListeners();
       this.linkEventListeners();
     }
-  
+
     addClasses(){
       this.svg.querySelector('style').appendChild(document.createTextNode(`
         .hover {transform: scale(1.05);transform-origin: 75% 75%;transition-duration:2s;}
@@ -24,7 +24,7 @@ class NavMap {
         .focus * {stroke-width: 6;stroke-miterlimit: 1;}
       `))
     }
-  
+
     addLinksToMap(){
       this.regions.forEach(region => {
         const thisCounterpart = this.mapCounterpart(region.id)
@@ -38,24 +38,24 @@ class NavMap {
         region.innerHTML = wrapLink
       })
     }
-  
+
     mapEventListeners() {
       this.regions.forEach(element => element.addEventListener("mouseenter", () => this.mapIn(element, "hover", true)))
       this.regions.forEach(element => element.addEventListener("mouseout", () => this.mapOut(element, "hover", true)))
       this.regions.forEach(element => element.addEventListener("click", event => this.mapClick(event)))
     }
-  
+
     linkEventListeners() {
       this.list.forEach(item => item.addEventListener("mouseenter", event => this.linkEvent(event.target, "in", "hover")))
       this.list.forEach(element => element.addEventListener("mouseout", event => this.linkEvent(event.target, "out", "hover")))
       this.list.forEach(item => item.addEventListener("focusin", event => this.linkEvent(event.target, "in", "focus")))
       this.list.forEach(item => item.addEventListener("focusout", event => this.linkEvent(event.target, "out", "focus")))
     }
-  
+
     moveToTop(obj) {
       obj.parentElement.appendChild(obj);
     }
-  
+
     mapIn(target, style, map) {
       this.moveToTop(target)
       target.classList.add(style)
@@ -64,7 +64,7 @@ class NavMap {
         if(thisLink) thisLink.classList.add("hover")
       }
     }
-  
+
     mapOut(target, style, map) {
       target.classList.remove(style)
       if(map) {
@@ -72,18 +72,18 @@ class NavMap {
         if(thisLink) thisLink.classList.remove("hover")
       }
     }
-  
+
     mapClick(event) {
       event.preventDefault()
       const thisMapCounterpart = this.mapCounterpart(event.target.closest("g").id)
       if(thisMapCounterpart) thisMapCounterpart.click()
     }
-  
+
     mapCounterpart(thisId) {
       const thisCounterpart = this.list.find(item => item.id === thisId)
       return (thisCounterpart)
     }
-  
+
     linkEvent(target, direction, type) {
       const thisId = target.id
       const thisCounterpart = this.linkCounterpart(thisId)
@@ -93,13 +93,13 @@ class NavMap {
         this.mapOut(thisCounterpart, type)
       }
     }
-  
+
     linkCounterpart(thisId) {
       const thisCounterpart = this.regions.find(item => item.id === thisId)
       return (thisCounterpart)
     }
   }
-  
+
   [...document.querySelectorAll('.nhsuk-map')].forEach(map => {
     // need to wait for SVG to load
     const obj = map.querySelector('object')
@@ -110,3 +110,4 @@ class NavMap {
       }
     })
   })
+}
