+++ app/assets/javascript/polyfills.js
@@ -1,0 +1,1 @@
+// Add polyfills (in case if any)
